New challenge: Fibonacci Order
Correct implementation for Fibonacci Order

- New algorithm taking advantage of the new restriction of the input
to lists of positive integers
- Make `SumSlice` objects immutable to better suit the new algorithm
- Add input validation and change documentation and tests
to take into account the new input restriction

New challenge (Fibonacci Order): first attempt

The implementation has a flaw that I discovered in the process of
writing an example runthrough. I'm not sure it can be easily patched
without affecting the time complexity of the algorithm,
so I'll be adding a constraint on the input for the problem --
requiring the integers in the input list to be strictly positive,
just like how it is in the original Code Golf Stack Exchange challenge.